break out LedgerController to avoid application fail when no Fabric service configured
