+++ src/main/java/org/sysc4907/votingsystem/Elections/ElectionController.java
@@ -36,10 +36,6 @@ public class ElectionController {
         this.electionService = electionService;
     }
 
-    @Autowired
-    private FabricGatewayService fabricGatewayService;
-
-
     @GetMapping("/create-election")
     public String showElectionForm(Model model) {
         model.addAttribute("electionForm", new ElectionForm());
@@ -86,111 +82,6 @@ public class ElectionController {
             model.addAttribute("errorMessage", "No poll has been configured yet!");
         }
         return "election-details";
-    }
-
-    @GetMapping("/ledger")
-    public String showElectionLedger(Model model, HttpSession session) {
-
-        /* Preserve election details here */
-
-        String username = (String) session.getAttribute("username");
-
-        model.addAttribute("isLoggedIn", username != null);
-        model.addAttribute("username", username);
-
-        LocalDateTime now = LocalDateTime.now();
-
-        // Pull this out for scope
-        Election election;
-
-        if (electionService.electionIsConfigured()) {
-            election = electionService.getElection();
-            model.addAttribute("election", election);
-
-            model.addAttribute("formattedStart", election.START_DATE_TIME.format(DateTimeFormatter.ofPattern("MMMM d, yyyy @ h:mm a")));
-            model.addAttribute("formattedEnd", election.END_DATE_TIME.format(DateTimeFormatter.ofPattern("MMMM d, yyyy @ h:mm a")));
-            model.addAttribute("currentCountdown", election.getElectionCountdown());
-            model.addAttribute("postElection", election.END_DATE_TIME.isBefore(now));
-
-            System.out.println(model.getAttribute("isLoggedIn"));
-            System.out.println(election.END_DATE_TIME);
-
-        } else {
-            model.addAttribute("errorMessage", "No poll has been configured yet!");
-        }
-
-        /* GET the Ledger and Tabulate */
-
-        // String holding all contents of ledger (JSON)
-        String ledgerString;
-        // Strings for call to FabricGatewayService
-        String function = "GetAllBallots";
-        String[] args = {};
-
-        // Retrieve ledger string from HyperLedger
-        try {
-            ledgerString = fabricGatewayService.evaluateTransaction(function, args);
-        } catch (Exception e) {
-            return "Error: " + e.getMessage();
-        }
-        model.addAttribute("ledgerString", ledgerString);
-
-        // Convert the JSON string to a JSONArray for easier manipulation
-        JSONArray ledgerJsonArray = new JSONArray(ledgerString);
-        // Container to hold LedgerRecord objects for pass to model
-        ArrayList<LedgerEntry> ledgerEntries = new ArrayList<>();
-
-        // Populate
-        for (int i = 0; i < ledgerJsonArray.length(); i++) {
-            JSONObject ledgerTransaction = (JSONObject) ledgerJsonArray.get(i);
-            // If valid number of fields, cast as LedgerRecord and add to array
-            if (ledgerTransaction.length() == 4) {
-                LedgerEntry lr = new LedgerEntry(
-                        (String) ledgerTransaction.getString("ballotId"),
-                        (String) ledgerTransaction.getString("ballotMarks"),
-                        (String) ledgerTransaction.getString("candidateOrder"),
-                        (String) ledgerTransaction.getString("ring"));
-                ledgerEntries.add(lr);
-            }
-        }
-
-        // Now, have an ArrayList of LedgerRecords we can pass to model
-        model.addAttribute("ledgerEntries", ledgerEntries);
-
-        /* File for Download */
-
-        return "ledger-all-votes";
-    }
-
-    // Example endpoint for downloading a string as a JSON file
-    @GetMapping("/download-ledger")
-    public ResponseEntity<String> downloadJsonFile(Model model, HttpSession session) {
-
-        /* GET the Ledger as a String */
-
-        // String holding all contents of ledger (JSON)
-        String ledgerString;
-
-        // Strings for call to FabricGatewayService
-        String function = "GetAllBallots";
-        String[] args = {};
-
-        // Retrieve ledger string from HyperLedger
-        try {
-            ledgerString = fabricGatewayService.evaluateTransaction(function, args);
-        } catch (Exception e) {
-            return new ResponseEntity<>("Error: Failed to retrieve JSON content.", HttpStatus.BAD_REQUEST);
-        }
-        model.addAttribute("ledgerString", ledgerString);
-
-        // Set headers for the file download
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Disposition", "attachment; filename=ledger.json");
-        headers.add("Content-Type", "application/json");
-
-        // Return the ResponseEntity with the JSON string
-        ResponseEntity responseEntity = new ResponseEntity<>(ledgerString, headers, HttpStatus.OK);
-        return responseEntity;
     }
 
 }
